Un-nest metadata file to network
