--- packages/lodestar/src/network/metadata/index.ts
@@ -1,1 +1,0 @@
-export * from "./metadata";
